Refactoring: unify creation of components
